+++ Src/Alco/Serialization/Binary/ReferenceContextExtension.cs
@@ -5,7 +5,7 @@ public static class ReferenceContextExtension
     public static void TryReadReferenceId(this ReferenceContext context, BinarySerializeReadNode node, ISerializable value)
     {
         if (value is IReferenceable referenceable
-            && node.Content.TryGetValue(ReferenceContext.SerializeKey, out uint id)
+            && node.Content.TryGetValue(ReferenceContext.SerializeKey, out ulong id)
             && id != 0)
         {
             context.SetReference(id, referenceable);
@@ -16,7 +16,7 @@ public static class ReferenceContextExtension
     {
         if (value is IReferenceable referenceable)
         {
-            uint id = context.GetId(referenceable);
+            ulong id = context.GetId(referenceable);
             node.SetValue(ReferenceContext.SerializeKey, id);
         }
     }
